Split virtual repo into its own module
